backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
boto3==1.40.26
PyJWT==2.10.1
passlib==1.7.4
bcrypt==3.2.0
email-validator==2.3.0
pytest==8.4.2
pytest-cov==6.2.1
httpx==0.28.1
ruff==0.7.0
